fixed configs

+++ src/tasks/services/tasks.service.ts
@@ -14,12 +14,12 @@ export class TasksService {
     private taskRepository: TaskRepository
   ) {}
   
-  public getTask(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  public getTask(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  public async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  public async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!found) {
       throw new NotFoundException(`Task with ID: ${id} not found`);
     }
@@ -33,15 +33,15 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  public async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  public async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with id: ${id} not found!`);
     }
   }
 
-  public async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.taskRepository.findOne(id);
+  public async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
